Update Rock Paper Scissor Shoot
Honestly I need to get better at this.

I want to add a score feature in this game, but I am stuck at one thing.

When I am referencing to variables in a method they do not come along with the variable I already defined in he main method. 

I want to know how to carry a variable across the main method.